requirements.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jsonapi==7.1.0
inflection==0.5.1
joblib==1.4.2
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tzdata==2025.1
